+++ TesteImposto/Imposto.Core.Data/NotaFiscalRepository.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Imposto.Core.Data
+{
+    public class NotaFiscalRepository
+    {
+
+    }
+}
