+++ TesteImposto/Imposto.Core.Service/XMLService.cs
@@ -1,0 +1,20 @@
+﻿using Imposto.Core.Domain;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Xml.Serialization;
+
+namespace Imposto.Core.Service
+{
+    public class XMLService
+    {   
+        public void GerarXML(Domain.Pedido pedido)
+        {
+            NotaFiscal notaFiscal = new NotaFiscal();
+            notaFiscal.GerarXML(pedido);
+        }
+    }
+}
