--- TesteImposto/Imposto.Core/Service/NotaFiscalService.cs
@@ -1,18 +1,0 @@
-﻿using Imposto.Core.Domain;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Imposto.Core.Service
-{
-    public class NotaFiscalService
-    {
-        public void GerarNotaFiscal(Domain.Pedido pedido)
-        {
-            NotaFiscal notaFiscal = new NotaFiscal();
-            notaFiscal.EmitirNotaFiscal(pedido);
-        }
-    }
-}
