+++ TesteImposto/SQL/CRIAR_BANCO_DADOS.sql
@@ -36,6 +36,10 @@ CREATE TABLE [dbo].[NotaFiscalItem](
 	[ValorIcms] [decimal](18, 5) NULL,
 	[NomeProduto] [varchar](50) NULL,
 	[CodigoProduto] [varchar](20) NULL,
+	[BaseIpi] [decimal](18, 5) NULL,
+	[AliquotaIpi] [decimal](18, 5) NULL,
+	[ValorIpi] [decimal](18, 5) NULL,	
+	[Desconto] [decimal](18, 5) NULL,
  CONSTRAINT [PK_NotaFiscalItem] PRIMARY KEY CLUSTERED 
 (
 	[Id] ASC
